requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
pandas==2.2.3
pkgconfig==1.5.5
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.13.1
